Fix things after rebase
