+++ src/pkg/services/m365/users.go
@@ -7,7 +7,7 @@ import (
 	"github.com/microsoftgraph/msgraph-sdk-go/models"
 
 	"github.com/alcionai/corso/src/internal/common/ptr"
-	"github.com/alcionai/corso/src/internal/m365/graph"
+	commonM365 "github.com/alcionai/corso/src/internal/m365"
 	"github.com/alcionai/corso/src/pkg/account"
 	"github.com/alcionai/corso/src/pkg/fault"
 	"github.com/alcionai/corso/src/pkg/path"
@@ -66,16 +66,7 @@ func UserHasMailbox(ctx context.Context, acct account.Account, userID string) (b
 		return false, clues.Stack(err).WithClues(ctx)
 	}
 
-	_, err = ac.Users().GetMailInbox(ctx, userID)
-	if err != nil {
-		if err := api.EvaluateMailboxError(err); err != nil {
-			return false, clues.Stack(err)
-		}
-
-		return false, nil
-	}
-
-	return true, nil
+	return commonM365.IsExchangeServiceEnabled(ctx, ac.Users(), userID)
 }
 
 // UserHasDrives returns true if the user has any drives
@@ -86,26 +77,7 @@ func UserHasDrives(ctx context.Context, acct account.Account, userID string) (bo
 		return false, clues.Stack(err).WithClues(ctx)
 	}
 
-	return checkUserHasDrives(ctx, ac.Users(), userID)
-}
-
-func checkUserHasDrives(ctx context.Context, dgdd getDefaultDriver, userID string) (bool, error) {
-	_, err := dgdd.GetDefaultDrive(ctx, userID)
-	if err != nil {
-		// we consider this a non-error case, since it
-		// answers the question the caller is asking.
-		if clues.HasLabel(err, graph.LabelsMysiteNotFound) || clues.HasLabel(err, graph.LabelsNoSharePointLicense) {
-			return false, nil
-		}
-
-		if graph.IsErrUserNotFound(err) {
-			return false, clues.Stack(graph.ErrResourceOwnerNotFound, err)
-		}
-
-		return false, clues.Stack(err)
-	}
-
-	return true, nil
+	return commonM365.IsOneDriveServiceEnabled(ctx, ac.Users(), userID)
 }
 
 // usersNoInfo returns a list of users in the specified M365 tenant - with no info
@@ -192,6 +164,8 @@ func parseUser(item models.Userable) (*User, error) {
 }
 
 // UserInfo returns the corso-specific set of user metadata.
+// TODO(pandeyabs): Remove support for this API. SDK users would be using
+// per service API calls - UserHasMailbox, UserGetMailboxInfo, UserHasDrive, etc.
 func GetUserInfo(
 	ctx context.Context,
 	acct account.Account,
@@ -208,4 +182,23 @@ func GetUserInfo(
 	}
 
 	return ui, nil
+}
+
+// TODO(pandeyabs): Add tests for this
+func UserGetMailboxInfo(
+	ctx context.Context,
+	acct account.Account,
+	userID string,
+) (api.MailboxInfo, error) {
+	ac, err := makeAC(ctx, acct, path.ExchangeService)
+	if err != nil {
+		return api.MailboxInfo{}, clues.Stack(err).WithClues(ctx)
+	}
+
+	mi, err := ac.Users().GetMailboxInfo(ctx, userID)
+	if err != nil {
+		return api.MailboxInfo{}, clues.Stack(err)
+	}
+
+	return mi, nil
 }
